Various documentation updates, including a modified main page and the removal of various obsolete platforms and modules.

+++ os/services/unit-test/unit-test.c
@@ -50,10 +50,10 @@ unit_test_print_report(const unit_test_t *utp)
   printf("\nUnit test: %s\n", utp->descr);
   printf("Result: %s\n", utp->result == unit_test_failure ?
                          "failure" : "success");
-  printf("Exit point: %s:%u\n", utp->test_file, utp->exit_line);
-  printf("Start: %lu\n", (unsigned long)utp->start);
-  printf("End: %lu\n", (unsigned long)utp->end);
-  printf("Duration: %lu\n", (unsigned long)(utp->end - utp->start));
-  printf("Ticks per second: %lu\n", (unsigned long)RTIMER_SECOND);
+  printf("Exit point: %s:%u\n", utp->test_file, (unsigned)utp->exit_line);
+  printf("Start: %u\n", (unsigned)utp->start);
+  printf("End: %u\n", (unsigned)utp->end);
+  printf("Duration: %u\n", (unsigned)(utp->end - utp->start));
+  printf("Ticks per second: %u\n", (unsigned)RTIMER_SECOND);
 }
 /*---------------------------------------------------------------------------*/
